requirements.txt
streamlit
vnstock
pandas
plotly
google-generativeai
